third_party/WebKit/LayoutTests/editing/selection/extend/extend_selection_word_13.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../assert_selection.js"></script><script>
function testIt(dir, offset, direction, expectedText) {
  test(() => assert_selection(
  `<span dir="${dir}">${makeSample(offset)}</span>`,
  selection => selection.modify('extend', direction, 'word'),
  `<span dir="${dir}">${expectedText}</span>`),
  `${dir} ${offset} extend ${direction} word`
  );
}

function makeSample(offset){
  const kSample = 'Lorem <div  dir="ltr">ipsum dolor sit</div> amet';
  return kSample.substr(0, offset) + '|' + kSample.substring(offset);
}

// ltr right word
testIt('ltr', 0, 'right', '^Lorem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 1, 'right', 'L^orem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 2, 'right', 'Lo^rem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 3, 'right', 'Lor^em| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 4, 'right', 'Lore^m| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 5, 'right', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('ltr', 6, 'right', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('ltr', 7, 'right', 'Lorem <^div|  dir="ltr">ipsum dolor sit amet');
testIt('ltr', 8, 'right', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('ltr', 9, 'right', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('ltr', 10, 'right', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('ltr', 11, 'right', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 12, 'right', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 13, 'right', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 14, 'right', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 15, 'right', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 16, 'right', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('ltr', 17, 'right', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('ltr', 18, 'right', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('ltr', 19, 'right', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('ltr', 20, 'right', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('ltr', 21, 'right', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 22, 'right', 'Lorem <div dir="ltr">^ipsum| dolor sit</div> amet');
testIt('ltr', 23, 'right', 'Lorem <div dir="ltr">i^psum| dolor sit</div> amet');
testIt('ltr', 24, 'right', 'Lorem <div dir="ltr">ip^sum| dolor sit</div> amet');
testIt('ltr', 25, 'right', 'Lorem <div dir="ltr">ips^um| dolor sit</div> amet');
testIt('ltr', 26, 'right', 'Lorem <div dir="ltr">ipsu^m| dolor sit</div> amet');
testIt('ltr', 27, 'right', 'Lorem <div dir="ltr">ipsum^ dolor| sit</div> amet');
testIt('ltr', 28, 'right', 'Lorem <div dir="ltr">ipsum ^dolor| sit</div> amet');
testIt('ltr', 29, 'right', 'Lorem <div dir="ltr">ipsum d^olor| sit</div> amet');
testIt('ltr', 30, 'right', 'Lorem <div dir="ltr">ipsum do^lor| sit</div> amet');
testIt('ltr', 31, 'right', 'Lorem <div dir="ltr">ipsum dol^or| sit</div> amet');
testIt('ltr', 32, 'right', 'Lorem <div dir="ltr">ipsum dolo^r| sit</div> amet');
testIt('ltr', 33, 'right', 'Lorem <div dir="ltr">ipsum dolor^ sit|</div> amet');
testIt('ltr', 34, 'right', 'Lorem <div dir="ltr">ipsum dolor ^sit|</div> amet');
testIt('ltr', 35, 'right', 'Lorem <div dir="ltr">ipsum dolor s^it|</div> amet');
testIt('ltr', 36, 'right', 'Lorem <div dir="ltr">ipsum dolor si^t|</div> amet');
testIt('ltr', 37, 'right', 'Lorem <div dir="ltr">ipsum dolor sit^</div> amet|');
testIt('ltr', 38, 'right', 'Lorem <div dir="ltr">ipsum dolor sit<^/div|> amet</div>');
testIt('ltr', 39, 'right', 'Lorem <div dir="ltr">ipsum dolor sitdiv^ amet|</div>');
testIt('ltr', 40, 'right', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 41, 'right', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 42, 'right', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 43, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('ltr', 44, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('ltr', 45, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> a^met|');
testIt('ltr', 46, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> am^et|');
testIt('ltr', 47, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> ame^t|');

// ltr left word
testIt('ltr', 0, 'left', '|Lorem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 1, 'left', '|L^orem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 2, 'left', '|Lo^rem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 3, 'left', '|Lor^em <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 4, 'left', '|Lore^m <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 5, 'left', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 6, 'left', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 7, 'left', '|Lorem <^div  dir="ltr">ipsum dolor sit amet');
testIt('ltr', 8, 'left', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('ltr', 9, 'left', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('ltr', 10, 'left', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('ltr', 11, 'left', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 12, 'left', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 13, 'left', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 14, 'left', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 15, 'left', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 16, 'left', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('ltr', 17, 'left', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('ltr', 18, 'left', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('ltr', 19, 'left', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('ltr', 20, 'left', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('ltr', 21, 'left', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 22, 'left', '|Lorem <div dir="ltr">^ipsum dolor sit</div> amet');
testIt('ltr', 23, 'left', 'Lorem <div dir="ltr">|i^psum dolor sit</div> amet');
testIt('ltr', 24, 'left', 'Lorem <div dir="ltr">|ip^sum dolor sit</div> amet');
testIt('ltr', 25, 'left', 'Lorem <div dir="ltr">|ips^um dolor sit</div> amet');
testIt('ltr', 26, 'left', 'Lorem <div dir="ltr">|ipsu^m dolor sit</div> amet');
testIt('ltr', 27, 'left', 'Lorem <div dir="ltr">|ipsum^ dolor sit</div> amet');
testIt('ltr', 28, 'left', 'Lorem <div dir="ltr">|ipsum ^dolor sit</div> amet');
testIt('ltr', 29, 'left', 'Lorem <div dir="ltr">ipsum |d^olor sit</div> amet');
testIt('ltr', 30, 'left', 'Lorem <div dir="ltr">ipsum |do^lor sit</div> amet');
testIt('ltr', 31, 'left', 'Lorem <div dir="ltr">ipsum |dol^or sit</div> amet');
testIt('ltr', 32, 'left', 'Lorem <div dir="ltr">ipsum |dolo^r sit</div> amet');
testIt('ltr', 33, 'left', 'Lorem <div dir="ltr">ipsum |dolor^ sit</div> amet');
testIt('ltr', 34, 'left', 'Lorem <div dir="ltr">ipsum |dolor ^sit</div> amet');
testIt('ltr', 35, 'left', 'Lorem <div dir="ltr">ipsum dolor |s^it</div> amet');
testIt('ltr', 36, 'left', 'Lorem <div dir="ltr">ipsum dolor |si^t</div> amet');
testIt('ltr', 37, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit^</div> amet');
testIt('ltr', 38, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit<^/div> amet</div>');
testIt('ltr', 39, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit^div amet</div>');
testIt('ltr', 40, 'left', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 41, 'left', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 42, 'left', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 43, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('ltr', 44, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('ltr', 45, 'left', 'Lorem <div dir="ltr">ipsum dolor sit</div> |a^met');
testIt('ltr', 46, 'left', 'Lorem <div dir="ltr">ipsum dolor sit</div> |am^et');
testIt('ltr', 47, 'left', 'Lorem <div dir="ltr">ipsum dolor sit</div> |ame^t');

// ltr forward word
testIt('ltr', 0, 'forward', '^Lorem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 1, 'forward', 'L^orem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 2, 'forward', 'Lo^rem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 3, 'forward', 'Lor^em| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 4, 'forward', 'Lore^m| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 5, 'forward', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('ltr', 6, 'forward', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('ltr', 7, 'forward', 'Lorem <^div|  dir="ltr">ipsum dolor sit amet');
testIt('ltr', 8, 'forward', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('ltr', 9, 'forward', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('ltr', 10, 'forward', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('ltr', 11, 'forward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 12, 'forward', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 13, 'forward', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 14, 'forward', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 15, 'forward', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 16, 'forward', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('ltr', 17, 'forward', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('ltr', 18, 'forward', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('ltr', 19, 'forward', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('ltr', 20, 'forward', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('ltr', 21, 'forward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 22, 'forward', 'Lorem <div dir="ltr">^ipsum| dolor sit</div> amet');
testIt('ltr', 23, 'forward', 'Lorem <div dir="ltr">i^psum| dolor sit</div> amet');
testIt('ltr', 24, 'forward', 'Lorem <div dir="ltr">ip^sum| dolor sit</div> amet');
testIt('ltr', 25, 'forward', 'Lorem <div dir="ltr">ips^um| dolor sit</div> amet');
testIt('ltr', 26, 'forward', 'Lorem <div dir="ltr">ipsu^m| dolor sit</div> amet');
testIt('ltr', 27, 'forward', 'Lorem <div dir="ltr">ipsum^ dolor| sit</div> amet');
testIt('ltr', 28, 'forward', 'Lorem <div dir="ltr">ipsum ^dolor| sit</div> amet');
testIt('ltr', 29, 'forward', 'Lorem <div dir="ltr">ipsum d^olor| sit</div> amet');
testIt('ltr', 30, 'forward', 'Lorem <div dir="ltr">ipsum do^lor| sit</div> amet');
testIt('ltr', 31, 'forward', 'Lorem <div dir="ltr">ipsum dol^or| sit</div> amet');
testIt('ltr', 32, 'forward', 'Lorem <div dir="ltr">ipsum dolo^r| sit</div> amet');
testIt('ltr', 33, 'forward', 'Lorem <div dir="ltr">ipsum dolor^ sit|</div> amet');
testIt('ltr', 34, 'forward', 'Lorem <div dir="ltr">ipsum dolor ^sit|</div> amet');
testIt('ltr', 35, 'forward', 'Lorem <div dir="ltr">ipsum dolor s^it|</div> amet');
testIt('ltr', 36, 'forward', 'Lorem <div dir="ltr">ipsum dolor si^t|</div> amet');
testIt('ltr', 37, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit^</div> amet|');
testIt('ltr', 38, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit<^/div|> amet</div>');
testIt('ltr', 39, 'forward', 'Lorem <div dir="ltr">ipsum dolor sitdiv^ amet|</div>');
testIt('ltr', 40, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 41, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 42, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 43, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('ltr', 44, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('ltr', 45, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> a^met|');
testIt('ltr', 46, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> am^et|');
testIt('ltr', 47, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> ame^t|');

// ltr backward word
testIt('ltr', 0, 'backward', '|Lorem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 1, 'backward', '|L^orem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 2, 'backward', '|Lo^rem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 3, 'backward', '|Lor^em <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 4, 'backward', '|Lore^m <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 5, 'backward', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 6, 'backward', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 7, 'backward', '|Lorem <^div  dir="ltr">ipsum dolor sit amet');
testIt('ltr', 8, 'backward', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('ltr', 9, 'backward', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('ltr', 10, 'backward', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('ltr', 11, 'backward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 12, 'backward', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 13, 'backward', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 14, 'backward', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 15, 'backward', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 16, 'backward', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('ltr', 17, 'backward', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('ltr', 18, 'backward', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('ltr', 19, 'backward', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('ltr', 20, 'backward', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('ltr', 21, 'backward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('ltr', 22, 'backward', '|Lorem <div dir="ltr">^ipsum dolor sit</div> amet');
testIt('ltr', 23, 'backward', 'Lorem <div dir="ltr">|i^psum dolor sit</div> amet');
testIt('ltr', 24, 'backward', 'Lorem <div dir="ltr">|ip^sum dolor sit</div> amet');
testIt('ltr', 25, 'backward', 'Lorem <div dir="ltr">|ips^um dolor sit</div> amet');
testIt('ltr', 26, 'backward', 'Lorem <div dir="ltr">|ipsu^m dolor sit</div> amet');
testIt('ltr', 27, 'backward', 'Lorem <div dir="ltr">|ipsum^ dolor sit</div> amet');
testIt('ltr', 28, 'backward', 'Lorem <div dir="ltr">|ipsum ^dolor sit</div> amet');
testIt('ltr', 29, 'backward', 'Lorem <div dir="ltr">ipsum |d^olor sit</div> amet');
testIt('ltr', 30, 'backward', 'Lorem <div dir="ltr">ipsum |do^lor sit</div> amet');
testIt('ltr', 31, 'backward', 'Lorem <div dir="ltr">ipsum |dol^or sit</div> amet');
testIt('ltr', 32, 'backward', 'Lorem <div dir="ltr">ipsum |dolo^r sit</div> amet');
testIt('ltr', 33, 'backward', 'Lorem <div dir="ltr">ipsum |dolor^ sit</div> amet');
testIt('ltr', 34, 'backward', 'Lorem <div dir="ltr">ipsum |dolor ^sit</div> amet');
testIt('ltr', 35, 'backward', 'Lorem <div dir="ltr">ipsum dolor |s^it</div> amet');
testIt('ltr', 36, 'backward', 'Lorem <div dir="ltr">ipsum dolor |si^t</div> amet');
testIt('ltr', 37, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit^</div> amet');
testIt('ltr', 38, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit<^/div> amet</div>');
testIt('ltr', 39, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit^div amet</div>');
testIt('ltr', 40, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 41, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 42, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('ltr', 43, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('ltr', 44, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('ltr', 45, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit</div> |a^met');
testIt('ltr', 46, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit</div> |am^et');
testIt('ltr', 47, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit</div> |ame^t');

// rtl right word
testIt('rtl', 0, 'right', '^Lorem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 1, 'right', 'L^orem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 2, 'right', 'Lo^rem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 3, 'right', 'Lor^em| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 4, 'right', 'Lore^m| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 5, 'right', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('rtl', 6, 'right', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('rtl', 7, 'right', 'Lorem <^div|  dir="ltr">ipsum dolor sit amet');
testIt('rtl', 8, 'right', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('rtl', 9, 'right', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('rtl', 10, 'right', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('rtl', 11, 'right', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 12, 'right', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 13, 'right', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 14, 'right', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 15, 'right', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 16, 'right', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('rtl', 17, 'right', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('rtl', 18, 'right', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('rtl', 19, 'right', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('rtl', 20, 'right', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('rtl', 21, 'right', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 22, 'right', 'Lorem <div dir="ltr">^ipsum| dolor sit</div> amet');
testIt('rtl', 23, 'right', 'Lorem <div dir="ltr">i^psum| dolor sit</div> amet');
testIt('rtl', 24, 'right', 'Lorem <div dir="ltr">ip^sum| dolor sit</div> amet');
testIt('rtl', 25, 'right', 'Lorem <div dir="ltr">ips^um| dolor sit</div> amet');
testIt('rtl', 26, 'right', 'Lorem <div dir="ltr">ipsu^m| dolor sit</div> amet');
testIt('rtl', 27, 'right', 'Lorem <div dir="ltr">ipsum^ dolor| sit</div> amet');
testIt('rtl', 28, 'right', 'Lorem <div dir="ltr">ipsum ^dolor| sit</div> amet');
testIt('rtl', 29, 'right', 'Lorem <div dir="ltr">ipsum d^olor| sit</div> amet');
testIt('rtl', 30, 'right', 'Lorem <div dir="ltr">ipsum do^lor| sit</div> amet');
testIt('rtl', 31, 'right', 'Lorem <div dir="ltr">ipsum dol^or| sit</div> amet');
testIt('rtl', 32, 'right', 'Lorem <div dir="ltr">ipsum dolo^r| sit</div> amet');
testIt('rtl', 33, 'right', 'Lorem <div dir="ltr">ipsum dolor^ sit|</div> amet');
testIt('rtl', 34, 'right', 'Lorem <div dir="ltr">ipsum dolor ^sit|</div> amet');
testIt('rtl', 35, 'right', 'Lorem <div dir="ltr">ipsum dolor s^it|</div> amet');
testIt('rtl', 36, 'right', 'Lorem <div dir="ltr">ipsum dolor si^t|</div> amet');
testIt('rtl', 37, 'right', 'Lorem <div dir="ltr">ipsum dolor sit^</div> amet|');
testIt('rtl', 38, 'right', 'Lorem <div dir="ltr">ipsum dolor sit<^/div|> amet</div>');
testIt('rtl', 39, 'right', 'Lorem <div dir="ltr">ipsum dolor sitdiv^ amet|</div>');
testIt('rtl', 40, 'right', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 41, 'right', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 42, 'right', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 43, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('rtl', 44, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('rtl', 45, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> a^met|');
testIt('rtl', 46, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> am^et|');
testIt('rtl', 47, 'right', 'Lorem <div dir="ltr">ipsum dolor sit</div> ame^t|');

// rtl left word
testIt('rtl', 0, 'left', '|Lorem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 1, 'left', '|L^orem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 2, 'left', '|Lo^rem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 3, 'left', '|Lor^em <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 4, 'left', '|Lore^m <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 5, 'left', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 6, 'left', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 7, 'left', '|Lorem <^div  dir="ltr">ipsum dolor sit amet');
testIt('rtl', 8, 'left', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('rtl', 9, 'left', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('rtl', 10, 'left', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('rtl', 11, 'left', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 12, 'left', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 13, 'left', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 14, 'left', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 15, 'left', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 16, 'left', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('rtl', 17, 'left', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('rtl', 18, 'left', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('rtl', 19, 'left', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('rtl', 20, 'left', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('rtl', 21, 'left', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 22, 'left', '|Lorem <div dir="ltr">^ipsum dolor sit</div> amet');
testIt('rtl', 23, 'left', 'Lorem <div dir="ltr">|i^psum dolor sit</div> amet');
testIt('rtl', 24, 'left', 'Lorem <div dir="ltr">|ip^sum dolor sit</div> amet');
testIt('rtl', 25, 'left', 'Lorem <div dir="ltr">|ips^um dolor sit</div> amet');
testIt('rtl', 26, 'left', 'Lorem <div dir="ltr">|ipsu^m dolor sit</div> amet');
testIt('rtl', 27, 'left', 'Lorem <div dir="ltr">|ipsum^ dolor sit</div> amet');
testIt('rtl', 28, 'left', 'Lorem <div dir="ltr">|ipsum ^dolor sit</div> amet');
testIt('rtl', 29, 'left', 'Lorem <div dir="ltr">ipsum |d^olor sit</div> amet');
testIt('rtl', 30, 'left', 'Lorem <div dir="ltr">ipsum |do^lor sit</div> amet');
testIt('rtl', 31, 'left', 'Lorem <div dir="ltr">ipsum |dol^or sit</div> amet');
testIt('rtl', 32, 'left', 'Lorem <div dir="ltr">ipsum |dolo^r sit</div> amet');
testIt('rtl', 33, 'left', 'Lorem <div dir="ltr">ipsum |dolor^ sit</div> amet');
testIt('rtl', 34, 'left', 'Lorem <div dir="ltr">ipsum |dolor ^sit</div> amet');
testIt('rtl', 35, 'left', 'Lorem <div dir="ltr">ipsum dolor |s^it</div> amet');
testIt('rtl', 36, 'left', 'Lorem <div dir="ltr">ipsum dolor |si^t</div> amet');
testIt('rtl', 37, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit^</div> amet');
testIt('rtl', 38, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit<^/div> amet</div>');
testIt('rtl', 39, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit^div amet</div>');
testIt('rtl', 40, 'left', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 41, 'left', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 42, 'left', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 43, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('rtl', 44, 'left', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('rtl', 45, 'left', 'Lorem <div dir="ltr">ipsum dolor sit</div> |a^met');
testIt('rtl', 46, 'left', 'Lorem <div dir="ltr">ipsum dolor sit</div> |am^et');
testIt('rtl', 47, 'left', 'Lorem <div dir="ltr">ipsum dolor sit</div> |ame^t');

// rtl forward word
testIt('rtl', 0, 'forward', '^Lorem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 1, 'forward', 'L^orem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 2, 'forward', 'Lo^rem| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 3, 'forward', 'Lor^em| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 4, 'forward', 'Lore^m| <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 5, 'forward', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('rtl', 6, 'forward', 'Lorem ^<div dir="ltr">ipsum| dolor sit</div> amet');
testIt('rtl', 7, 'forward', 'Lorem <^div|  dir="ltr">ipsum dolor sit amet');
testIt('rtl', 8, 'forward', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('rtl', 9, 'forward', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('rtl', 10, 'forward', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('rtl', 11, 'forward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 12, 'forward', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 13, 'forward', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 14, 'forward', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 15, 'forward', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 16, 'forward', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('rtl', 17, 'forward', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('rtl', 18, 'forward', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('rtl', 19, 'forward', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('rtl', 20, 'forward', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('rtl', 21, 'forward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 22, 'forward', 'Lorem <div dir="ltr">^ipsum| dolor sit</div> amet');
testIt('rtl', 23, 'forward', 'Lorem <div dir="ltr">i^psum| dolor sit</div> amet');
testIt('rtl', 24, 'forward', 'Lorem <div dir="ltr">ip^sum| dolor sit</div> amet');
testIt('rtl', 25, 'forward', 'Lorem <div dir="ltr">ips^um| dolor sit</div> amet');
testIt('rtl', 26, 'forward', 'Lorem <div dir="ltr">ipsu^m| dolor sit</div> amet');
testIt('rtl', 27, 'forward', 'Lorem <div dir="ltr">ipsum^ dolor| sit</div> amet');
testIt('rtl', 28, 'forward', 'Lorem <div dir="ltr">ipsum ^dolor| sit</div> amet');
testIt('rtl', 29, 'forward', 'Lorem <div dir="ltr">ipsum d^olor| sit</div> amet');
testIt('rtl', 30, 'forward', 'Lorem <div dir="ltr">ipsum do^lor| sit</div> amet');
testIt('rtl', 31, 'forward', 'Lorem <div dir="ltr">ipsum dol^or| sit</div> amet');
testIt('rtl', 32, 'forward', 'Lorem <div dir="ltr">ipsum dolo^r| sit</div> amet');
testIt('rtl', 33, 'forward', 'Lorem <div dir="ltr">ipsum dolor^ sit|</div> amet');
testIt('rtl', 34, 'forward', 'Lorem <div dir="ltr">ipsum dolor ^sit|</div> amet');
testIt('rtl', 35, 'forward', 'Lorem <div dir="ltr">ipsum dolor s^it|</div> amet');
testIt('rtl', 36, 'forward', 'Lorem <div dir="ltr">ipsum dolor si^t|</div> amet');
testIt('rtl', 37, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit^</div> amet|');
testIt('rtl', 38, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit<^/div|> amet</div>');
testIt('rtl', 39, 'forward', 'Lorem <div dir="ltr">ipsum dolor sitdiv^ amet|</div>');
testIt('rtl', 40, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 41, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 42, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 43, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('rtl', 44, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> ^amet|');
testIt('rtl', 45, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> a^met|');
testIt('rtl', 46, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> am^et|');
testIt('rtl', 47, 'forward', 'Lorem <div dir="ltr">ipsum dolor sit</div> ame^t|');

// rtl backward word
testIt('rtl', 0, 'backward', '|Lorem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 1, 'backward', '|L^orem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 2, 'backward', '|Lo^rem <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 3, 'backward', '|Lor^em <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 4, 'backward', '|Lore^m <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 5, 'backward', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 6, 'backward', '|Lorem^ <div dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 7, 'backward', '|Lorem <^div  dir="ltr">ipsum dolor sit amet');
testIt('rtl', 8, 'backward', 'Lorem <d|iv dir="ltr">ipsum dolor sit amet</d|iv>');
testIt('rtl', 9, 'backward', 'Lorem <di|v dir="ltr">ipsum dolor sit amet</di|v>');
testIt('rtl', 10, 'backward', 'Lorem <div| dir="ltr">ipsum dolor sit amet</div|>');
testIt('rtl', 11, 'backward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 12, 'backward', 'Lorem <div |dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 13, 'backward', 'Lorem <div d|ir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 14, 'backward', 'Lorem <div di|r="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 15, 'backward', 'Lorem <div dir|="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 16, 'backward', 'Lorem <div dir="|&quot;ltr&quot;">ipsum dolor sit</div> amet');
testIt('rtl', 17, 'backward', 'Lorem <div dir="|ltr">ipsum dolor sit</div> amet');
testIt('rtl', 18, 'backward', 'Lorem <div dir="l|tr">ipsum dolor sit</div> amet');
testIt('rtl', 19, 'backward', 'Lorem <div dir="lt|r">ipsum dolor sit</div> amet');
testIt('rtl', 20, 'backward', 'Lorem <div dir="ltr|">ipsum dolor sit</div> amet');
testIt('rtl', 21, 'backward', 'Lorem <div | dir="ltr">ipsum dolor sit</div> amet');
testIt('rtl', 22, 'backward', '|Lorem <div dir="ltr">^ipsum dolor sit</div> amet');
testIt('rtl', 23, 'backward', 'Lorem <div dir="ltr">|i^psum dolor sit</div> amet');
testIt('rtl', 24, 'backward', 'Lorem <div dir="ltr">|ip^sum dolor sit</div> amet');
testIt('rtl', 25, 'backward', 'Lorem <div dir="ltr">|ips^um dolor sit</div> amet');
testIt('rtl', 26, 'backward', 'Lorem <div dir="ltr">|ipsu^m dolor sit</div> amet');
testIt('rtl', 27, 'backward', 'Lorem <div dir="ltr">|ipsum^ dolor sit</div> amet');
testIt('rtl', 28, 'backward', 'Lorem <div dir="ltr">|ipsum ^dolor sit</div> amet');
testIt('rtl', 29, 'backward', 'Lorem <div dir="ltr">ipsum |d^olor sit</div> amet');
testIt('rtl', 30, 'backward', 'Lorem <div dir="ltr">ipsum |do^lor sit</div> amet');
testIt('rtl', 31, 'backward', 'Lorem <div dir="ltr">ipsum |dol^or sit</div> amet');
testIt('rtl', 32, 'backward', 'Lorem <div dir="ltr">ipsum |dolo^r sit</div> amet');
testIt('rtl', 33, 'backward', 'Lorem <div dir="ltr">ipsum |dolor^ sit</div> amet');
testIt('rtl', 34, 'backward', 'Lorem <div dir="ltr">ipsum |dolor ^sit</div> amet');
testIt('rtl', 35, 'backward', 'Lorem <div dir="ltr">ipsum dolor |s^it</div> amet');
testIt('rtl', 36, 'backward', 'Lorem <div dir="ltr">ipsum dolor |si^t</div> amet');
testIt('rtl', 37, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit^</div> amet');
testIt('rtl', 38, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit<^/div> amet</div>');
testIt('rtl', 39, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit^div amet</div>');
testIt('rtl', 40, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 41, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 42, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit amet</div>');
testIt('rtl', 43, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('rtl', 44, 'backward', 'Lorem <div dir="ltr">ipsum dolor |sit</div>^ amet');
testIt('rtl', 45, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit</div> |a^met');
testIt('rtl', 46, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit</div> |am^et');
testIt('rtl', 47, 'backward', 'Lorem <div dir="ltr">ipsum dolor sit</div> |ame^t');
</script>
